chore: Move iterators to their own module.
